requirements.txt
pandas==2.2.3
numpy==2.0.2
scipy==1.13.1
requests==2.32.3
pyarrow==19.0.1
python-dotenv==1.0.1
rich==13.7.1
flask==3.1.3
gunicorn==23.0.0
